requirements.txt
evdev
python-rtmidi